+++ Chapter.Net.WPF.Converters/EqualsToBooleanConverter/EqualsToBooleanConverter.cs
@@ -7,6 +7,7 @@
 using System;
 using System.ComponentModel;
 using System.Globalization;
+using System.Linq;
 using System.Windows.Data;
 
 // ReSharper disable once CheckNamespace
@@ -17,6 +18,7 @@ namespace Chapter.Net.WPF.Converters;
 ///     Compares a single value to a variable or a list of values to each other and returns its boolean representation.
 /// </summary>
 [ValueConversion(typeof(object), typeof(bool))]
+[ValueConversion(typeof(object[]), typeof(bool))]
 public class EqualsToBooleanConverter : SingleAndMultiValueConverter
 {
     /// <summary>
@@ -54,16 +56,22 @@ public class EqualsToBooleanConverter : SingleAndMultiValueConverter
     }
 
     /// <summary>
-    ///     ConvertBack is not implemented.
+    ///     Compares a list of values to each other and returns its boolean representation.
     /// </summary>
-    /// <param name="value">Unused.</param>
+    /// <param name="values">The values to convert.</param>
     /// <param name="targetType">Unused.</param>
     /// <param name="parameter">Unused.</param>
     /// <param name="culture">Unused.</param>
-    /// <returns>Nothing.</returns>
-    /// <exception cref="NotImplementedException">The EqualsToBooleanConverter.ConvertBack is not implemented.</exception>
-    public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+    /// <returns>The converted value.</returns>
+    public override object Convert(object[] values, Type targetType, object parameter, CultureInfo culture)
     {
-        throw new NotImplementedException();
+        if (values == null)
+            return IsNotEqual;
+
+        if (values.Length == 0)
+            return IsNotEqual;
+
+        var allEqual = values.All(o => Equals(o, values[0]));
+        return allEqual ? IsEqual : IsNotEqual;
     }
 }